**EDIT: I also found the bug that was causing us to OOM when executing instant queries, see the 2nd commit**

This should reduce the memory allocated by the vector iteration by at least 25% but as much as ~80%. The quantile vector iteration accounts for at least 15% of all in use space last time I tested (and during that time queriers we're OOM'ing as a result of quantile sharding). 

However, with the usage of this alternate store type we give up some amount of relative accuracy guarantees for _lower_ quantiles. From their readme:

```
We also provide implementations, returned by LogCollapsingLowestDenseDDSketch(relativeAccuracy, maxNumBins) and LogCollapsingHighestDenseDDSketch(relativeAccuracy, maxNumBins), where the q-quantile will be accurate up to the specified relative error for q that is not too small (or large). Concretely, the q-quantile will be accurate up to the specified relative error as long as it belongs to one of the m bins kept by the sketch. For instance, If the values are time in seconds, maxNumBins = 2048 covers a time range from 80 microseconds to 1 year.
```

Benchmark:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/logql
cpu: AMD Ryzen 9 5950X 16-Core Processor
                                                      │    base     │                new2                │
                                                      │   sec/op    │   sec/op     vs base               │
QuantileBatchRangeVectorIteratorAt/1-samples-32         153.4n ± 1%   169.8n ± 1%  +10.76% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/1000-samples-32      30.17µ ± 1%   10.63µ ± 1%  -64.76% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/10000-samples-32     281.1µ ± 1%   101.5µ ± 2%  -63.88% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/100000-samples-32    2.355m ± 2%   1.006m ± 2%  -57.26% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/1000000-samples-32   23.50m ± 2%   10.11m ± 1%  -56.99% (p=0.002 n=6)
geomean                                                 148.4µ        71.47µ       -51.84%

                                                      │     base     │                 new2                 │
                                                      │     B/op     │    B/op      vs base                 │
QuantileBatchRangeVectorIteratorAt/1-samples-32          80.00 ±  0%   80.00 ±  0%        ~ (p=1.000 n=6) ¹
QuantileBatchRangeVectorIteratorAt/1000-samples-32      157.00 ±  1%   80.00 ±  0%  -49.04% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/10000-samples-32     155.00 ±  3%   80.00 ±  1%  -48.39% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/100000-samples-32    109.00 ± 76%   83.00 ±  0%  -23.85% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/1000000-samples-32   1082.5 ±  4%   222.0 ± 11%  -79.49% (p=0.002 n=6)
geomean                                                  187.2         98.84        -47.19%
¹ all samples are equal

                                                      │    base    │                new2                 │
                                                      │ allocs/op  │ allocs/op   vs base                 │
QuantileBatchRangeVectorIteratorAt/1-samples-32         3.000 ± 0%   3.000 ± 0%        ~ (p=1.000 n=6) ¹
QuantileBatchRangeVectorIteratorAt/1000-samples-32      6.000 ± 0%   3.000 ± 0%  -50.00% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/10000-samples-32     6.000 ± 0%   3.000 ± 0%  -50.00% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/100000-samples-32    4.000 ± 0%   3.000 ± 0%  -25.00% (p=0.002 n=6)
QuantileBatchRangeVectorIteratorAt/1000000-samples-32   7.000 ± 0%   3.000 ± 0%  -57.14% (p=0.002 n=6)
geomean                                                 4.967        3.000       -39.60%
¹ all samples are equal
```

And profiling
![2024-02-08-180659_2040x719_scrot](https://github.com/grafana/loki/assets/3246492/89770447-8eb1-4914-96d5-b702e06bd6d4)
![2024-02-08-180725_2042x650_scrot](https://github.com/grafana/loki/assets/3246492/3d9cf1c4-6eac-4c31-a3b5-3b38e613b2f9)